feat: Add basic testing (MPT-4914)

+++ app/schemas.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import datetime
 import uuid
-from decimal import Decimal
 from typing import Annotated
 
 from pydantic import BaseModel, ConfigDict, Field
@@ -128,5 +127,5 @@ class DataSourceRead(BaseSchema):
     organization_id: uuid.UUID
     type: DataSourceType
     resources_changed_this_month: int
-    expenses_so_far_this_month: Decimal
-    expenses_forecast_this_month: Decimal
+    expenses_so_far_this_month: float
+    expenses_forecast_this_month: float
